feat: support --no-clean arg for build

+++ src/scripts/build/rollup.js
@@ -43,7 +43,9 @@ const scripts = args.includes('--p-react')
   ? getPReactScripts()
   : getConcurrentlyArgs(getCommands())
 
-rimraf.sync(fromRoot('dist'))
+if (!args.includes('--no-clean')) {
+  rimraf.sync(fromRoot('dist'))
+}
 
 const result = spawn.sync(resolveBin('concurrently'), scripts, {
   stdio: 'inherit',
